fix bugs and move ensureColorScheme to fiftyone/state
