+++ app/packages/app/src/useEvents/useSetColorScheme.ts
@@ -1,13 +1,20 @@
-import { useSessionSetter } from "@fiftyone/state";
+import { ensureColorScheme, useSessionSetter } from "@fiftyone/state";
 import { useCallback } from "react";
 import { EventHandlerHook } from "./registerEvent";
-import { ensureColorScheme } from "./utils";
 
 const useSetColorScheme: EventHandlerHook = () => {
   const setter = useSessionSetter();
   return useCallback(
     (payload) => {
-      setter("colorScheme", ensureColorScheme(payload.color_scheme));
+      setter(
+        "colorScheme",
+        ensureColorScheme(payload.color_scheme, {
+          colorPool: [],
+          colorBy: "field",
+          opacity: 0.7,
+          multicolorKeypoints: false,
+        })
+      );
     },
     [setter]
   );
